third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/navigating-across-documents/014-expected.txt

This is a testharness.js-based test.
FAIL  Link with javascript onclick form submission script order  assert_array_equals: property 0, expected "after script" but got "submit"
Harness: the test ran to completion.
